FileHandler.handle

+++ models/fileHandler.js
@@ -13,6 +13,34 @@ class FileHandler {
   }
 
   /**
+   * Handles file uploads.
+   * @param files {{}} - An object containing uploaded files.
+   * @returns {Promise<FileHandler>} - A Promise that resolves with a new
+   *   FileHandler instance for the uploaded files.
+   */
+
+  static async handle (files) {
+    if (files.file && files.file.mimetype && files.file.size) {
+      const { size } = files.file
+      const mime = files.file.mimetype
+      const now = new Date()
+      const timestamp = now.getTime() / 1000
+      let name, thumbnail
+
+      if (mime.startsWith('image/')) {
+        const keys = await FileHandler.handleArt(files.file, files.thumbnail)
+        name = keys.file
+        thumbnail = keys.thumbnail
+      } else {
+        const res = await FileHandler.upload(files.file)
+        name = res.key
+      }
+
+      if (name) return new FileHandler({ name, thumbnail, mime, size, timestamp })
+    }
+  }
+
+  /**
    * Handles uploading art.
    * @param art {!{ name: string, data: Buffer, mimetype: string }} - The image
    *   to upload.
@@ -67,9 +95,11 @@ class FileHandler {
    */
 
   static async remove (key) {
-    const s3 = FileHandler.instantiateS3()
-    const params = {Bucket: config.aws.bucket, Key: key}
-    await s3.deleteObject(params).promise()
+    if (key) {
+      const s3 = FileHandler.instantiateS3()
+      const params = {Bucket: config.aws.bucket, Key: key}
+      await s3.deleteObject(params).promise()
+    }
   }
 
   /**
